添加后台请求地址访问权限，页面功能权限

+++ src/main/java/com/ramostear/jbuilder/controller/ArticleController.java
@@ -9,6 +9,7 @@
 								                 limitations under the License. 
 */
 package com.ramostear.jbuilder.controller;
+import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -38,12 +39,14 @@ public class ArticleController {
 	@Autowired
 	private AttachmentService attachmentService;
 	
+	@RequiresPermissions(value="article")
 	@RequestMapping(value="/index",method=RequestMethod.GET)
 	public String index(Model model){
 		model.addAttribute("statusList",articleService.getStatusCount());
 		return "article/index";
 	}
 	
+	@RequiresPermissions(value="article:list")
 	@RequestMapping(value="/list",method=RequestMethod.GET)
 	public String list(Article article,ReqDto req,Model model,String search){
 		model.addAttribute("list", articleService.findByPage(req.getPageNo(), req.getPageSize(), "createTime", true,search));
@@ -51,30 +54,32 @@ public class ArticleController {
 		return "article/list";
 	}
 	
+	@RequiresPermissions(value="article:add")
 	@RequestMapping(value="/add",method=RequestMethod.GET)
 	public String add(Model m){
 		return "article/add";
 	}
 	
-	
+	@RequiresPermissions(value="article:add")
 	@RequestMapping(value="/add",method=RequestMethod.POST)
 	public String add(Article article,MultipartFile file,String status){
 		if(file!=null){
 			String url = QiniuFileUtil.upload(file);
 			article.setCover(url);
 		}
-		System.out.println(status);
 		article.setStatus(Integer.parseInt(status));
 		articleService.add(article);
 		return "article/index";
 	}
 	
+	@RequiresPermissions(value="article:edit")
 	@RequestMapping(value="/edit",method=RequestMethod.GET)
 	public String eidt(Model m,Long id){
 		m.addAttribute("article",this.articleService.findById(id));
 		return "article/add";
 	}
 	
+	@RequiresPermissions(value="article:edit")
 	@RequestMapping(value="/edit",method=RequestMethod.POST)
 	public String eidt(Article article,MultipartFile file,String status){
 		String url="";
@@ -92,9 +97,14 @@ public class ArticleController {
 		return "article/index";
 	}
 	
+	@RequiresPermissions(value="article:delete")
 	@RequestMapping(value="/delete",method=RequestMethod.GET)
 	public String delete(Long id){
-		articleService.delete(id);
+		Article temp=this.articleService.findById(id);
+		if(temp!=null){
+			QiniuFileUtil.deleteQiniuFile(temp.getCover());
+			articleService.delete(id);
+		}
 		return "article/index";
 	}
 }
